Add support for <M-BS> as delete-backward-word
Although <C-W> is perfectly useful already, one of the readline key
bindings I use a lot is <M-BS> ("delete-backward-word").  However,
implementing this is complicated by the fact that it is apparently
impossible(?) to get the actual backspace key within vim.  Attempts to,
for example,

	exe "set <F34>=\<Esc>\<BS>"

result in <F34> being set to <Esc><80>fb, which is probably some
internal indirect reference to the termcap variable, and, moreover,
doesn't work. Inexplicably, as well, one cannot retrieve the value of
&t_xx (&t_fb specifically is what we want here) termcap options in vim
script.

As far as I know, in the modern era, terminals generally use either
<C-H> or <C-?> for backspace; and sometimes these are transposed with
the bindings for <Del>. My brutish solution is to bind both <Esc><C-H>
and <Esc><C-?> to the desired function. The consequence is that, for
some terminals, <M-Del> might be bound to the same as <M-BS>, but
I don't think <M-Del> means anything in readline anyhow.