Support meta maps with modifyOtherKeys
Some terminals, including xterm and iTerm, contain support for alternate
escape sequences for meta keys presses.  For example, Meta-a, normally
encoded as "<Esc>a", is instead encoded as "<Esc>[27;3;97~", where "3"
means "meta", and "97" is the ASCII decimal representation of "a".  Our
":set" hack won't catch these; we need the native "M-" map.

The rub is that Vim internally represents meta key combos as the regular
ASCII character with the 8th bit set.  For example, since ASCII "a" is
0x61, <M-a> is 0xE1, which in UTF-8 is encoded as 0xC3 0x81.  But in
Unicode, "á" is also character U+00E1, which means any map of <M-a> will
also trigger upon typing "á".

However, if we define the map in another encoding, the key combo gets
encoded as a different sequence of bytes.  For example, in the latin1
encoding, we get the simple 1 byte sequence of 0xE1, which doesn't
conflict with "á", but does conflict with, among other characters,
U+1E05, aka "ḅ", which in UTF-8 is encoded as 0xE1 0xB8 0x85.

So what's a locale that won't give us a conflict?  If we start with the
list at `:help encoding-values` and eliminate anything 8-bit or Unicode
derived, we're left with a handful of options, most of which are flagged
as "Windows only" or "Unix only".  The Korean encoding cp949 is listed
as supporting both platforms, and encodes <M-a> as 0x80 0xFC 0x08 0x65.
This is not a valid UTF-8 sequence, even if we start mid-character, so
it should serve our needs perfectly.

Shout out to Neovim for doing what Vim should have done 15 years ago and
fixing this positively archaic quirk.